+++ src/main/java/lab1/tbd/serviciovoluntariado/repositories/EmergenciaRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 public interface EmergenciaRepository {
     public String createEmergencia(Emergencia emergencia); //Create
     public List<Emergencia> getAllEmergencias(); //Read
-    public String updateEmergencia(int id, Emergencia emergencia); //Update
-    public String deleteEmergencia(int id);
-    public Emergencia getEmergenciaById(Integer id);
-    public int getIdEmergenciaMayor();
+    public String updateEmergencia(Long id, Emergencia emergencia); //Update
+    public String deleteEmergencia(Long id);
+    public Emergencia getEmergenciaById(Long id);
+    public Long getIdEmergenciaMayor();
 }
