+++ src/main/java/lab1/tbd/serviciovoluntariado/repositories/EquipamientoRepository.java
@@ -8,9 +8,9 @@ import java.util.List;
 public interface EquipamientoRepository {
 
     public String createEquipamiento(Equipamiento equipamiento);
-    public int getIdEquipamientoMayor();
+    public Long getIdEquipamientoMayor();
     public List<Equipamiento> getAllEquipamiento();
-    public Equipamiento getEquipamientoById(int id);
-    public String updateEquipamiento(int id,Equipamiento equipamiento);
-    public String deleteEquipamiento(int id);
+    public Equipamiento getEquipamientoById(Long id);
+    public String updateEquipamiento(Long id,Equipamiento equipamiento);
+    public String deleteEquipamiento(Long id);
 }
