+++ src/main/java/lab1/tbd/serviciovoluntariado/repositories/VolHabilidadRepository.java
@@ -8,8 +8,8 @@ public interface VolHabilidadRepository {
     
     public int countVolHabilidad();
     public List<VolHabilidad> getAllVolHabilidad();
-    public VolHabilidad getVolHabilidadById(int id);
+    public VolHabilidad getVolHabilidadById(Long id);
     public VolHabilidad createVolHabilidad(VolHabilidad v);
     public VolHabilidad updateVolHabilidad(VolHabilidad v);
-    public void deleteVolHabilidadById(int id);
+    public void deleteVolHabilidadById(Long id);
 }
