+++ src/main/java/lab1/tbd/serviciovoluntariado/repositories/VolParticipacionRepository.java
@@ -3,5 +3,5 @@ import lab1.tbd.serviciovoluntariado.models.VolParticipacion;
 import java.util.List;
 
 public interface VolParticipacionRepository {
-    public List<VolParticipacion> getVoluntariesPerParticipation(int id);
+    public List<VolParticipacion> getVoluntariesPerParticipation(Long id);
 }
