+++ src/main/java/lab1/tbd/serviciovoluntariado/repositories/VolRankingRequerimientosRepository.java
@@ -3,5 +3,5 @@ import lab1.tbd.serviciovoluntariado.models.VolCantidadHabilidades;
 import java.util.List;
 
 public interface VolRankingRequerimientosRepository {
-    public List<VolCantidadHabilidades> getVoluntariosRequerimiento(int idEmergencia);
+    public List<VolCantidadHabilidades> getVoluntariosRequerimiento(Long idEmergencia);
 }
