+++ src/main/java/lab1/tbd/serviciovoluntariado/repositories/VoluntarioRepository.java
@@ -6,8 +6,8 @@ import java.util.List;
 public interface VoluntarioRepository{
     public int countVoluntario();
     public List<Voluntario> getAllVoluntario();
-    public Voluntario getVoluntarioById(int id);
+    public Voluntario getVoluntarioById(Long id);
     public Voluntario createVoluntario(Voluntario v);
     public Voluntario updateVoluntario(Voluntario v);
-    public void deleteVoluntarioById(int id);
+    public void deleteVoluntarioById(Long id);
 }
